% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gp.R
\name{gp_setup}
\alias{gp_setup}
\title{gp}
\usage{
gp_setup(xObs, yObs, xPred)
}
\arguments{
\item{xObs}{observed x values}

\item{yObs}{observed y values}

\item{xPred}{predicted x values}
}
\value{
a list object containing handles to the compiled MCMC routine, prediction routine, and nimble model object
}
\description{
gp
}

